requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
DateTime==5.5
fake-useragent==1.5.1
feedparser==6.0.11
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
lxml==5.3.0
nltk==3.9.1
numpy==2.1.3
pandas==2.2.3
parse==1.20.2
praw==7.8.1
prawcore==2.4.0
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
textblob==0.18.0.post0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==1.26.20
w3lib==2.2.1
websocket-client==1.8.0
websockets==10.4
yahoo-fin==0.8.9.1
zipp==3.21.0
zope.interface==7.2
